requirements.txt
Django==3.2.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
python-decouple==3.5
psycopg2-binary==2.9.3
django-filter==21.1
gunicorn==20.1.0
dj-database-url==0.5.0
whitenoise==5.3.0
factory-boy==3.2.1
Faker==11.3.0
django-nose==1.4.7
coverage==6.2
dropbox==11.25.0
django-storages==1.12.3
Pillow==9.0.0
drf-yasg==1.20.0
django-cors-headers==3.11.0
drf-extra-fields==3.2.1





